component directory structure refactoring
